pruatlas: A package for PRU maps generation.

The pruatlas package provides function to plot country FIRs, access the underlying FIR or country spatial definitions.

Functions

The pruatlas package provides three sets of functions, one for plottting, the second for accessing the underlying datasets and the last is about helper functions. For example plot_coutry_fir is an high level function that plots the FIR at the country level.

TODOs

list/define the colours of the maps (currently hardcoded)

provide FIRs for other AIRAC's (separate data packages? lazy-loading of data?)

define a function to convert from CFMU AIRAC format to AIP format

change `airacnm` to the AIP format